% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_git.R
\name{check_git}
\alias{check_git}
\title{Check that \code{git} is installed}
\usage{
check_git()
}
\value{
Nothing.
}
\description{
Check that \code{git} is installed
}
\examples{
\dontrun{
  check_git()
}
}
\author{
Richèl J.C. Bilderbeek
}
