% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linux.R
\name{check_linux}
\alias{check_linux}
\title{Check that this package is run under Linux}
\usage{
check_linux()
}
\value{
Nothing.
}
\description{
Check that this package is run under Linux
}
\examples{
check_linux()
}
\author{
Richèl J.C. Bilderbeek
}
